+++ lib/model/middle-layer/src/pframe/internal_api/api_factory.ts
@@ -1,5 +1,4 @@
 import {
-  Base64Encoded,
   BinaryPartitionedDataInfo,
   JsonDataInfo,
   JsonPartitionedDataInfo,