+++ etc/blocks/monetization-test/model/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @milaboratories/milaboratories.monetization-test.model
 
+## 1.0.28
+
+### Patch Changes
+
+- 5814e48: Small changes in sdk
+- Updated dependencies [55b218b]
+- Updated dependencies [5814e48]
+  - @platforma-sdk/model@1.45.26
+
 ## 1.0.27
 
 ### Patch Changes