+++ etc/blocks/ui-examples/block/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @milaboratories/milaboratories.ui-examples
 
+## 1.0.430
+
+### Patch Changes
+
+- 5814e48: Small changes in sdk
+- Updated dependencies [55b218b]
+- Updated dependencies [5814e48]
+  - @platforma-sdk/model@1.45.26
+  - @milaboratories/milaboratories.ui-examples.workflow@1.1.101
+  - @milaboratories/milaboratories.ui-examples.model@1.3.3
+  - @milaboratories/milaboratories.ui-examples.ui@1.6.3
+
 ## 1.0.429
 
 ### Patch Changes