+++ lib/model/middle-layer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @milaboratories/pl-model-middle-layer
 
+## 1.8.36
+
+### Patch Changes
+
+- 55b218b: PFrameDriver decomposed
+- Updated dependencies [55b218b]
+  - @milaboratories/pl-model-common@1.21.5
+
 ## 1.8.35
 
 ### Patch Changes