+++ lib/node/ts-helpers-winston/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @milaboratories/ts-helpers-winston
 
+## 1.0.31
+
+### Patch Changes
+
+- Updated dependencies [55b218b]
+  - @milaboratories/ts-helpers@1.5.2
+
 ## 1.0.30
 
 ### Patch Changes